Refactors the code in the singlton design pattern
The original code was a little misleading and it also would compile with
a runtime error.

This offers a new version that illistrates the use of closures inside a
class definition. It also provides an example of how to handle modules
(including private classes) The two example together illistrate some of
the flexability and elegance CoffeeScript can offer. It also explains
how the use of the wrapper around a CS file.