Levenshtein, programming style
The version before this commit was I believe correct.
I would therefore totally understand if this pull request was rejected.

However I think levenshtein provides a very nice example to demonstrate
some of the functionality and pitfalls of coffeescript.

First I am not too fond of the "return unless" statement. It
feels a bit uncanny to find a non indented return in the middle
of a function.

Most important I am not fond of what this statement is trying to hide. In
most languages, checking for the empty string is not useful. 

see Discussion section.