Changed to best practice of the `this` keyword
I've seen several people on IRC talk about readability. It seems there is some opinions that when chaining like this the `this` keyword is more readable then a dangling `@` symbol.

I think it is a good idea to offer code readability in this project.

Feel free to revert this commit if this is disputed. (Unable to open a PR from in the Github UI).